livewell-e/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.42.0
langgraph==0.2.38
typing-extensions>=4.12
langchain-core>=0.2.39
tenacity>=9.0.0
httpx>=0.27
structlog>=24.1
